Add files via upload
Reimplemented [CSVState] as a dataviewtype instead of as a viewtype. This increased boilerplate but meant nicer notation in the main function.
Changed so that the delimiter and the option to allow quoted newline characters are now passed by templates.
Extended the mains function to be more illustrative.
Added more comments to the code.